+++ src/rics/misc.py
@@ -351,6 +351,10 @@ def _get_module(arg: type[_t.Any] | _t.Any) -> str:
 def format_kwargs(kwargs: _t.Mapping[str, _t.Any], *, max_value_length: int = 80) -> str:
     """Format keyword arguments.
 
+    .. warning::
+
+       This function will be removed in a future version. Use :func:`.strings.format_kwargs` instead.
+
     Args:
         kwargs: Arguments to format.
         max_value_length: Replace value with class name above this limit. 0=no limit.
@@ -367,7 +371,8 @@ def format_kwargs(kwargs: _t.Mapping[str, _t.Any], *, max_value_length: int = 80
         >>> format_kwargs({"an_int": 1, "a_string": "Hello!"})
         "an_int=1, a_string='Hello!'"
     """
-    # TODO(0.7.0): max_value_length=0
+    # TODO(0.7.0): deprecate in favor of string.format_kwargs() + update docstring ::warning
+
     invalid = [k for k in kwargs if not k.isidentifier()]
     if invalid:
         raise ValueError(f"Got {len(invalid)} invalid identifiers: {invalid}.")